첫 화면 만들기 하단 바 프래그먼트 적용 InitActivity 통해 초기 액티비티 설정 뒤 상속
