--- app/src/main/java/com/example/myapplication/ui/bottomBar/BottomBarFragment.java
@@ -1,26 +1,0 @@
-package com.example.myapplication.ui.bottomBar;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
-
-import com.example.myapplication.R;
-
-public class BottomBarFragment extends Fragment {
-    private BottomBarFragment( ) { }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        return inflater.inflate(R.layout.fragment_bottom_bar_open, container, false);
-    }
-}
