+++ app/src/main/res/layout/fragment_bottom_bar_close.xml
@@ -3,6 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/fragment_bottom_bar_close"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#00000000"
