+++ app/src/main/res/layout/fragment_bottom_bar_open.xml
@@ -2,23 +2,23 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_bottom_bar"
+    android:id="@+id/fragment_bottom_bar_open"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#22000000"
+    android:background="#33000000"
     android:orientation="vertical">
 
     <View
-        android:id="@+id/close_bottom_bar"
+        android:id="@+id/view_close_bottom_bar"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3"></View>
+        android:layout_height="400dp"
+        android:layout_weight="1"></View>
 
     <LinearLayout
         android:id="@+id/bottom_bar"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="2"
+        android:layout_height="240dp"
+        android:layout_weight="1"
         android:background="#9CACFF"
         android:gravity="center"
         android:orientation="vertical">
@@ -26,8 +26,8 @@
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="20dp"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:padding="20dp">
 
             <FrameLayout
                 android:layout_width="wrap_content"
@@ -51,14 +51,14 @@
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
-                android:text="asdfasdfasfd"/>
+                android:text="asdfasdfasfd" />
         </LinearLayout>
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="20dp"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:padding="20dp">
 
             <FrameLayout
                 android:layout_width="wrap_content"
@@ -75,14 +75,15 @@
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
-                android:text="asdfasdfasfd"/>
+                android:text="asdfasdfasfd" />
         </LinearLayout>
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:padding="20dp"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:padding="20dp">
+
             <FrameLayout
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
@@ -95,10 +96,11 @@
                     android:src="@drawable/icon3"
                     android:visibility="visible" />
             </FrameLayout>
+
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
-                android:text="asdfasdfasfd"/>
+                android:text="asdfasdfasfd" />
         </LinearLayout>
     </LinearLayout>
 </LinearLayout>
